raise_error.py is moved into testgres/operations from testgres/helpers
Let's store our things on one place.

We use RaiseError only in testgres/operations structures currently.
